// frontend/src/utils/types.ts
export interface customer_size_data_type {
  main_width: number;
  pants_width: number;
  shoulder_width: number;
  stein_width: number;
  stein_radius: number;
  side_hight: number;
  throat_radius: number;
  pants_radius: number;
  stein_model: string;
  collar_model: string;
}

export interface customer_data_type {
  id: number;
  user_name: string;
  last_name: string;
  phone: string;
}